third_party/WebKit/LayoutTests/web-animations-api/animations-responsive-to-inherited-change-expected.txt

This is a testharness.js-based test.
FAIL Font size responsive to inherited changes assert_equals: expected "25px" but got "15px"
Harness: the test ran to completion.
